pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-binary-bot"
version = "0.2.0"
description = "Deterministic binary-market bot scaffold with SQLite truth and replay-driven paper execution."
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
